pyproject.toml
[project]
name = "what_can_we_play"
version = "0.3.4"
description = "Something something"
authors = [{name = "dunderrrrrr", email = "emil.bjorkroth@gmail.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "pre-commit>=3.7.1",
    "pytest>=8.2.2",
    "mypy>=1.15.0",
    "htpy>=25.10.0",
    "flask>=3.1.2",
    "python-dotenv>=1.1.1",
    "python3-openid>=3.2.0",
    "flask-openid>=1.3.1",
    "python-steam-api>=2.2.1",
    "flask-caching>=2.3.1",
]


[tool.pytest.ini_options]
testpaths = [
    "tests/*",
]
